+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 * Filter by pending requests and resourceType
 * Assert by pending requests and resourceType
+* Browser.elementHtml and assertions
 
 2.1.3/ 2018-06-21
 =================